+++ src/Type/Square/Aztec/Encode.php
@@ -32,301 +32,27 @@ use Com\Tecnick\Barcode\Exception as BarcodeException;
  * @license     http://www.gnu.org/copyleft/lesser.html GNU-LGPL v3 (see LICENSE.TXT)
  * @link        https://github.com/tecnickcom/tc-lib-barcode
  */
-abstract class Encode extends \Com\Tecnick\Barcode\Type\Square\Aztec\Layers
+class Encode extends \Com\Tecnick\Barcode\Type\Square\Aztec\Bitstream
 {
      /**
-     * Returns the high-level encoding for the given code and ECI mode.
+     * Aztec main encoder.
      *
      * @param string $code The code to encode.
+     * @param int $ecc The error correction code percentage of error check words.
      * @param int $eci The ECI mode to use.
+     * @param string $hint The mode to use.
      */
-    protected function highLevelEncoding($code, $eci = 0, $hint = 'A')
+    public function __construct($code, $ecc = 33, $eci = 0, $hint = 'A')
     {
-        $this->addFLG($eci);
-        $chars = array_values(unpack('C*', $code));
-        $chrlen = count($chars);
-        if ($chrlen == 0) {
-            return;
+        $this->highLevelEncoding($code, $eci, $hint);
+        if ($this->totbits == 0) {
+              throw new BarcodeException('No input data');
         }
-        if ($hint == 'B') {
-            $this->binaryEncode($chars, $chrlen);
-            return;
+        if (!$this->sizeAndBitStuffing($ecc)) {
+                throw new BarcodeException('Data too long');
         }
-        $this->autoEncode($chars, $chrlen);
-    }
-
-    /**
-     * Forced binary encoding for the given characters.
-     *
-     * @param array $chars  Integer ASCII values of the characters to encode.
-     * @param int   $chrlen Lenght of the $chars array.
-     */
-    protected function binaryEncode($chars, $chrlen)
-    {
-        $bits = Data::MODE_BITS[Data::MODE_BINARY];
-        $this->addShift(Data::MODE_BINARY);
-        if ($chrlen > 62) {
-            $this->addRawCwd(5, 0);
-            $this->addRawCwd(11, $chrlen);
-            for ($idx = 0; $idx < $chrlen; $idx++) {
-                $this->addRawCwd($bits, $chars[$idx]);
-            }
-            return;
-        }
-        if ($chrlen > 31) {
-            $this->addRawCwd(5, 31);
-            for ($idx = 0; $idx < 31; $idx++) {
-                $this->addRawCwd($bits, $chars[$idx]);
-            }
-            $this->addShift(Data::MODE_BINARY);
-            $this->addRawCwd(5, ($chrlen - 31));
-            for ($idx = 31; $idx < $chrlen; $idx++) {
-                $this->addRawCwd($bits, $chars[$idx]);
-            }
-            return;
-        }
-        $this->addRawCwd(5, $chrlen);
-        for ($idx = 0; $idx < $chrlen; $idx++) {
-            $this->addRawCwd($bits, $chars[$idx]);
-        }
-    }
-
-    /**
-     * Automatic encoding for the given characters.
-     *
-     * @param array $chars  Integer ASCII values of the characters to encode.
-     * @param int   $chrlen Lenght of the $chars array.
-     */
-    protected function autoEncode($chars, $chrlen)
-    {
-        for ($idx = 0; $idx < $chrlen; $idx++) {
-            if ($this->processBinaryChars($chars, $idx, $chrlen)) {
-                continue;
-            }
-            if ($this->processPunctPairs($chars, $idx, $chrlen)) {
-                continue;
-            }
-            $this->processModeChars($chars, $idx, $chrlen);
-        }
-    }
-
-    /**
-     * Process mode characters.
-     *
-     * @param array $chars The array of characters.
-     * @param int $idx The current character index.
-     * @param int $chrlen The total numebr of characters to process.
-     */
-    protected function processModeChars(&$chars, &$idx, $chrlen)
-    {
-        $ord = $chars[$idx];
-        if ($this->isSameMode($this->encmode, $ord)) {
-            $mode = $this->encmode;
-        } else {
-            $mode = $this->charMode($ord);
-        }
-        $nchr = $this->countModeChars($chars, $idx, $chrlen, $mode);
-        if ($this->encmode != $mode) {
-            if (($nchr == 1) && (!empty(Data::SHIFT_MAP[$this->encmode][$mode]))) {
-                $this->addShift($mode);
-            } else {
-                $this->addLatch($mode);
-            }
-        }
-        $this->mergeTmpCwd();
-    }
-
-    /**
-    * Count consecutive characters in the same mode.
-    *
-    * @param array $chars The array of characters.
-    * @param int $idx The current character index.
-    * @param int $chrlen The total numebr of characters to process.
-    * @param int $mode The current mode.
-    *
-    * @return int
-    */
-    protected function countModeChars(&$chars, $idx, $chrlen, $mode)
-    {
-        $this->tmpCdws = array();
-        $nbits = Data::MODE_BITS[$mode];
-        $count = 0;
-        do {
-            $ord = $chars[$idx];
-            if (!$this->isSameMode($mode, $ord)) {
-                return $count;
-            }
-            $this->tmpCdws[] = array($nbits, $this->charEnc($mode, $ord));
-            $count++;
-            $idx++;
-        } while ($idx < $chrlen);
-        return $count;
-    }
-
-    /**
-     * Process consecutive binary characters.
-     *
-     * @param array $chars The array of characters.
-     * @param int $idx The current character index.
-     * @param int $chrlen The total numebr of characters to process.
-     *
-     * @return bool True if binary characters have been found and processed.
-     */
-    protected function processBinaryChars(&$chars, &$idx, $chrlen)
-    {
-        $binchrs = $this->countBinaryChars($chars, $idx, $chrlen);
-        if ($binchrs == 0) {
-            return false;
-        }
-        $this->addShift(Data::MODE_BINARY);
-        if ($binchrs > 62) {
-            $this->addRawCwd(5, 0);
-            $this->addRawCwd(11, $binchrs);
-            $this->mergeTmpCwdRaw();
-            return true;
-        }
-        if ($binchrs > 31) {
-            $nbits = Data::MODE_BITS[Data::MODE_BINARY];
-            $this->addRawCwd(5, 31);
-            for ($bcw = 0; $bcw < 31; $bcw++) {
-                $this->addRawCwd($nbits, $this->tmpCdws[$bcw][1]);
-            }
-            $this->addShift(Data::MODE_BINARY);
-            $this->addRawCwd(5, ($binchrs - 31));
-            for ($bcw = 31; $bcw < $binchrs; $bcw++) {
-                $this->addRawCwd($nbits, $this->tmpCdws[$bcw][1]);
-            }
-            return true;
-        }
-        $this->addRawCwd(5, $binchrs);
-        $this->mergeTmpCwdRaw();
-        return true;
-    }
-
-    /**
-    * Count consecutive binary characters.
-    *
-    * @param array $chars The array of characters.
-    * @param int $idx The current character index.
-    * @param int $chrlen The total numebr of characters to process.
-    *
-    * @return int
-    *
-    * @SuppressWarnings(PHPMD.CyclomaticComplexity)
-    */
-    protected function countBinaryChars(&$chars, $idx, $chrlen)
-    {
-        $this->tmpCdws = array();
-        $count = 0;
-        $nbits = Data::MODE_BITS[Data::MODE_BINARY];
-        while (($idx < $chrlen) && ($count < 2048)) {
-            $ord = $chars[$idx];
-            if ($this->charMode($ord) != Data::MODE_BINARY) {
-                return $count;
-            }
-            $this->tmpCdws[] = array($nbits, $ord);
-            $count++;
-            $idx++;
-        }
-        return $count;
-    }
-
-    /**
-     * Process consecutive special Punctuation Pairs.
-     *
-     * @param array $chars The array of characters.
-     * @param int $idx The current character index.
-     * @param int $chrlen The total numebr of characters to process.
-     *
-     * @return bool True if pair characters have been found and processed.
-     *
-     * @SuppressWarnings(PHPMD.CyclomaticComplexity)
-     */
-    protected function processPunctPairs(&$chars, &$idx, $chrlen)
-    {
-        $ppairs = $this->countPunctPairs($chars, $idx, $chrlen);
-        if ($ppairs == 0) {
-            return false;
-        }
-        switch ($this->encmode) {
-            case Data::MODE_PUNCT:
-                break;
-            case Data::MODE_MIXED:
-                $this->addLatch(Data::MODE_PUNCT);
-                break;
-            case Data::MODE_UPPER:
-            case Data::MODE_LOWER:
-                if ($ppairs > 1) {
-                    $this->addLatch(Data::MODE_PUNCT);
-                }
-                break;
-            case Data::MODE_DIGIT:
-                $common = $this->countPunctAndDigitChars($chars, $idx, $chrlen);
-                if ($common < 6) {
-                    $this->mergeTmpCwdRaw();
-                    $idx += $common;
-                    return true;
-                }
-                if ($ppairs > 2) {
-                    $this->addLatch(Data::MODE_PUNCT);
-                }
-                break;
-            default:
-                return false;
-        }
-        $this->mergeTmpCwd(Data::MODE_PUNCT);
-        $idx += ($ppairs * 2);
-        return true;
-    }
-
-    /**
-     * Count consecutive special Punctuation Pairs.
-     *
-     * @param array $chars The array of characters.
-     * @param int $idx The current character index.
-     * @param int $chrlen The total numebr of characters to process.
-     *
-     * @return int
-     */
-    protected function countPunctPairs(&$chars, $idx, $chrlen)
-    {
-        $this->tmpCdws = array();
-        $pairs = 0;
-        $maxidx = $chrlen - 1;
-        while ($idx < $maxidx) {
-            $pmode = $this->punctPairMode($chars[$idx], $chars[($idx + 1)]);
-            if ($pmode == 0) {
-                return $pairs;
-            }
-            $this->tmpCdws[] = array(5, $pmode);
-            $pairs++;
-            $idx += 2;
-        }
-        return $pairs;
-    }
-
-    /**
-     * Counts the number of consecutive charcters that are in both PUNCT or DIGIT modes.
-     *
-     * @param array &$chars The string to count the characters in.
-     * @param int $idx The starting index to count from.
-     * @param int $chrlen The length of the string to count.
-     *
-     * @return int The number of punctuation and digit characters in the string.
-     */
-    protected function countPunctAndDigitChars(&$chars, $idx, $chrlen)
-    {
-        $this->tmpCdws = array();
-        $count = 0;
-        while ($idx < $chrlen) {
-            $ord = $chars[$idx];
-            if (!$this->isPunctAndDigitChar($ord)) {
-                return $count;
-            }
-            $this->tmpCdws[] = array(4, $this->charEnc(Data::MODE_DIGIT, $ord));
-            $count++;
-            $idx++;
-        }
-        return $count;
+       // TODO:
+       //  - Appending check codewords.
+       //  - Arranging the complete message in a spiral around the core.
     }
 }
